refactor: move getPlayerList function to common utils and update imports -playerListをAPI介さず直接DBに取りに行く
