+++ app/(linkheader)/umpire/page.tsx
@@ -2,6 +2,8 @@ import { getUmpireWithCompetition, groupByUmpire } from "@/app/lib/db/queries/qu
 import type { SelectUmpire, SelectUmpireWithCompetition } from "@/app/lib/db/schema"
 import { View } from "@/app/components/common/view"
 
+export const revalidate = 0
+
 export default async function Player() {
   const initialUmpireDataList: SelectUmpireWithCompetition[] = await groupByUmpire(await getUmpireWithCompetition())
