+++ app/components/common/commonRegister.tsx
@@ -1,8 +1,8 @@
 "use client"
 import { useCallback, useState } from "react"
 import type { SelectPlayer, SelectUmpire, SelectCompetition } from "@/app/lib/db/schema"
-import { getPlayerList } from "@/app/components/challenge/utils"
-import { getUmpireList, getCompetitionList } from "@/app/components/common/utils"
+// import { getPlayerList } from "@/app/components/challenge/utils"
+import { getPlayerList, getUmpireList, getCompetitionList } from "@/app/components/common/utils"
 
 type CommonRegisterProps = {
   type: "player" | "umpire" | "course" | "competition"
